[Finishes #168895370] added first middleware and configured travis ci for the project
