+++ .travis.yml
@@ -1,8 +1,5 @@
 language: node_js
 node_js:
-  - "6"
-  - "5"
-  - "4"
-  - "0.12"
+  - "10.16.3"
 script: npm run coverage
-after_success: 'npm run coveralls'q
+after_success: npm run coveralls
