+++ server/controllers/userController.js
@@ -4,23 +4,7 @@ import '@babel/polyfill'
 
 const UserController = {
     async createNewuser(req, res) {
-        const schema = {
-            first_name: Joi.string().required(),
-            last_name: Joi.string().required(),
-            email: Joi.string().email().required(),
-            password: Joi.string().min(6).required(),
-            gender: Joi.string().required(),
-            job_title: Joi.string().required(),
-            department: Joi.string().required(),
-            address: Joi.string().required(),
-        }
-        const result = Joi.validate(req.body, schema);
-        if (result.error) {
-            return res.status(400).send({
-                "status": 400,
-                "error": result.error.details[0].message
-            });
-        }
+        
         try {
             const user = UserModel.addNewUser(req);
             return res.status(201).send({
@@ -63,7 +47,6 @@ const UserController = {
                 "data": login_user.data
             })
         }catch(e){
-            console.log(e)
             return res.status(500).send({
                 "status": 500,
                 "error": "server error"
